refactor: rename `Provider` to `StoreProvider`
* change `Provider` as named export
* extract `useTheme` hook and `ThemeContext`
