+++ apps/calendar/src/components/provider/theme.tsx
@@ -1,19 +1,6 @@
-import { h, createContext, FunctionComponent } from 'preact';
+import { FunctionComponent, h } from 'preact';
 import Theme from '@src/theme';
-import { useContext } from 'preact/hooks';
-
-// eslint-disable-next-line no-undefined
-export const ThemeContext = createContext<Theme | undefined>(undefined);
-
-export const useTheme = (): Theme => {
-  const ctx = useContext(ThemeContext);
-
-  if (typeof ctx === 'undefined') {
-    throw new Error('Theme is not found');
-  }
-
-  return ctx as Theme;
-};
+import { ThemeContext } from '@src/components/hooks/theme';
 
 export const ThemeProvider: FunctionComponent<{ theme: Theme }> = ({ theme, children }) => (
   <ThemeContext.Provider value={theme}>{children}</ThemeContext.Provider>
