+++ apps/calendar/src/components/view/Main.tsx
@@ -1,11 +1,13 @@
 import { h, FunctionComponent } from 'preact';
 
-import Provider from '@src/components/provider';
+import { StoreProvider } from '@src/components/provider/store';
 import Store from '@src/store';
 import { RouterContext, useCreateRouter } from '@src/components/hooks/router';
 import ToolbarContainer from '@src/components/toolbar/toolbarContainer';
 import ViewContainer from '@src/components/view/viewContainer';
 import { ViewListMap } from '@t/option';
+import { ThemeProvider } from '@src/components/provider/theme';
+import Theme from '@src/theme';
 
 type Props = {
   initialView: string;
@@ -13,15 +15,19 @@ type Props = {
   store: Store;
 };
 
+// @TODO: Main component should be able to share the store and the theme to the other sub-calendars(Month, Week, Day, etc.)
 const Main: FunctionComponent<Props> = ({ initialView, components, store }) => {
   const router = useCreateRouter({ initialView, components });
+  const theme = new Theme();
 
   return (
     <RouterContext.Provider value={router}>
       <ToolbarContainer />
-      <Provider store={store}>
-        <ViewContainer />
-      </Provider>
+      <ThemeProvider theme={theme}>
+        <StoreProvider store={store}>
+          <ViewContainer />
+        </StoreProvider>
+      </ThemeProvider>
     </RouterContext.Provider>
   );
 };
