+++ apps/calendar/test/helper/storeProvider.tsx
@@ -1,6 +1,6 @@
 import { h, FunctionComponent } from 'preact';
 import Store from '@src/store';
-import StoreProvider from '@src/components/provider';
+import { StoreProvider } from '@src/components/provider/store';
 
 export function getStoreWrapper(store: Store): FunctionComponent {
   return function wrapper({ children }) {
